Extracted Equation class (Calculator)
